Event Timing: event counting (https://w3c.github.io/event-timing)
Implements event counting (performance.eventCounts), which is part of
the event timing API.

This implementation is incomplete: after the first access to
performance.eventCounts, the JS-visible object will no longer be
updated, even though the underlying C++ counters are still being
incremented.

Reviewed by NOBODY (OOPS!).

Explanation of why this fixes the bug (OOPS!).

* LayoutTests/TestExpectations: skip tests that timeout instead of failing early
* LayoutTests/imported/w3c/web-platform-tests/event-timing/event-click-counts-expected.txt: now passes
* LayoutTests/imported/w3c/web-platform-tests/event-timing/event-counts-zero-expected.txt: now passes
* LayoutTests/imported/w3c/web-platform-tests/event-timing/idlharness.window-expected.txt: now passes
* Source/WebCore/page/PointerLockController.cpp: fix build
* Source/WebCore/platform/Logging.h: add a new logger for the Performance Timeline