Event Timing: event counting (https://w3c.github.io/event-timing) https://webkit.org/b/296383 rdar://156500312
Reviewed by NOBODY (OOPS!).

Implements event counting (performance.eventCounts), which is part of
the event timing API.

The way this code implements a maplike interface is problematic: when
the eventCounts object is accessed through javascript, a second
object containing the event counts is created, and must be kept in
sync after that. See `EventCounts::add()` and
`EventCounts::initializeMapLike()`.

This implementation does not count mouse{enter,leave} and
pointer{enter,leave} events if there are no PerformanceObservers
watching them. This happens because an optimization prevents these
events from being dispatched; see
`EventHandler::updateMouseEventTargetNode()`.

* LayoutTests/TestExpectations: skip tests that timeout instead of failing early
* LayoutTests/imported/w3c/web-platform-tests/event-timing/event-click-counts-expected.txt: now passes
* LayoutTests/imported/w3c/web-platform-tests/event-timing/event-counts-zero-expected.txt: now passes
* LayoutTests/imported/w3c/web-platform-tests/event-timing/idlharness.window-expected.txt: now passes
* Source/WebCore/page/PointerLockController.cpp: fix build
* Source/WebCore/platform/Logging.h: add a new logger for the Performance Timeline